// src/app/model/stock.ts
export class Stock {
  lastAccessTime?: string;
  yearHigh?: number;
  price?: number;
  yearLow?: number;
  open?: number;
  previousClose?: number;
  id?: number;
  userId?: number;
  stockName?: string;
  userNote?: string;
  todayGain?: number;
  lastWeekGain?: number;
  twoWeeksGain?: number;
  threeWeeksGain?: number;
  monthGain?: number;
  current?: number;
}